scripts/requirements.txt
requests>=2.28.0
Pillow>=9.0.0
pytest>=7.0.0
responses>=0.23.0
